Move logs options out of experimental namespace
