+++ sentry-samples/sentry-samples-spring-boot-jakarta-opentelemetry-noagent/src/main/resources/application.properties
@@ -15,7 +15,7 @@ sentry.graphql.ignored-error-types=SOME_ERROR,ANOTHER_ERROR
 sentry.enable-backpressure-handling=true
 sentry.enable-spotlight=true
 sentry.enablePrettySerializationOutput=false
-sentry.experimental.logs.enabled=true
+sentry.logs.enabled=true
 in-app-includes="io.sentry.samples"
 
 # Uncomment and set to true to enable aot compatibility
