+++ sentry-samples/sentry-samples-spring-boot-jakarta/src/main/resources/application.properties
@@ -16,7 +16,7 @@ sentry.enable-backpressure-handling=true
 sentry.enable-spotlight=true
 sentry.enablePrettySerializationOutput=false
 in-app-includes="io.sentry.samples"
-sentry.experimental.logs.enabled=true
+sentry.logs.enabled=true
 
 # Uncomment and set to true to enable aot compatibility
 # This flag disables all AOP related features (i.e. @SentryTransaction, @SentrySpan)
