+++ sentry-samples/sentry-samples-spring-boot-opentelemetry/src/main/resources/application.properties
@@ -14,7 +14,7 @@ sentry.debug=true
 sentry.graphql.ignored-error-types=SOME_ERROR,ANOTHER_ERROR
 sentry.enable-backpressure-handling=true
 sentry.enable-spotlight=true
-sentry.experimental.logs.enabled=true
+sentry.logs.enabled=true
 in-app-includes="io.sentry.samples"
 
 # Database configuration
