+++ sentry-samples/sentry-samples-spring-boot-webflux-jakarta/src/main/resources/application.properties
@@ -10,5 +10,5 @@ sentry.logging.minimum-breadcrumb-level=debug
 sentry.reactive.thread-local-accessor-enabled=true
 sentry.traces-sample-rate=1.0
 sentry.enable-backpressure-handling=true
-sentry.experimental.logs.enabled=true
+sentry.logs.enabled=true
 sentry.enable-spotlight=true
