+++ sentry-samples/sentry-samples-spring-boot-webflux/src/main/resources/application.properties
@@ -12,5 +12,5 @@ spring.graphql.graphiql.enabled=true
 spring.graphql.websocket.path=/graphql
 spring.graphql.schema.printer.enabled=true
 sentry.enable-backpressure-handling=true
-sentry.experimental.logs.enabled=true
+sentry.logs.enabled=true
 sentry.enable-spotlight=true
