+++ sentry-spring-boot-jakarta/src/test/kotlin/io/sentry/spring/boot/jakarta/SentryAutoConfigurationTest.kt
@@ -188,7 +188,7 @@ class SentryAutoConfigurationTest {
             "sentry.cron.default-timezone=America/New_York",
             "sentry.cron.default-failure-issue-threshold=40",
             "sentry.cron.default-recovery-threshold=50",
-            "sentry.experimental.logs.enabled=true"
+            "sentry.logs.enabled=true"
         ).run {
             val options = it.getBean(SentryProperties::class.java)
             assertThat(options.readTimeoutMillis).isEqualTo(10)
@@ -233,7 +233,7 @@ class SentryAutoConfigurationTest {
             assertThat(options.cron!!.defaultTimezone).isEqualTo("America/New_York")
             assertThat(options.cron!!.defaultFailureIssueThreshold).isEqualTo(40L)
             assertThat(options.cron!!.defaultRecoveryThreshold).isEqualTo(50L)
-            assertThat(options.experimental.logs.isEnabled).isEqualTo(true)
+            assertThat(options.logs.isEnabled).isEqualTo(true)
         }
     }
