+++ sentry/src/main/java/io/sentry/ExperimentalOptions.java
@@ -1,8 +1,6 @@
 package io.sentry;
 
 import io.sentry.protocol.SdkVersion;
-import org.jetbrains.annotations.ApiStatus;
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -12,17 +10,6 @@ import org.jetbrains.annotations.Nullable;
  * <p>Beware that experimental options can change at any time.
  */
 public final class ExperimentalOptions {
-  private @NotNull SentryOptions.Logs logs = new SentryOptions.Logs();
 
   public ExperimentalOptions(final boolean empty, final @Nullable SdkVersion sdkVersion) {}
-
-  @ApiStatus.Experimental
-  public @NotNull SentryOptions.Logs getLogs() {
-    return logs;
-  }
-
-  @ApiStatus.Experimental
-  public void setLogs(@NotNull SentryOptions.Logs logs) {
-    this.logs = logs;
-  }
 }
