+++ sentry/src/main/java/io/sentry/logger/LoggerApi.java
@@ -92,7 +92,7 @@ public final class LoggerApi implements ILoggerApi {
         return;
       }
 
-      if (!options.getExperimental().getLogs().isEnabled()) {
+      if (!options.getLogs().isEnabled()) {
         options
             .getLogger()
             .log(SentryLevel.WARNING, "Sentry Log is disabled and this 'logger' call is a no-op.");
